weechat-channel-server/pyproject.toml
[project]
name = "weechat-channel-server"
version = "0.1.0"
description = "Claude Code Channel plugin bridging Zenoh P2P messaging and Claude Code"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "eclipse-zenoh>=1.0.0",
]

[project.scripts]
weechat-channel = "server:main"
